+++ itest/test_wrap.py
@@ -46,7 +46,7 @@ class TestWrap(TestBase):
             t = {
                 people: set({name: 'Iris'})
             };
-            return(t.wrap('People'), 2);
+            return t.wrap('People'), 2;
         ''')
         self.assertEqual(res, {"people": [{"name": "Iris"}]})
 
@@ -54,7 +54,7 @@ class TestWrap(TestBase):
             t = PS{
                 people: set({name: 'Iris'})
             };
-            return(t.wrap('People'), 2);
+            return t.wrap('People'), 2;
         ''')
         self.assertEqual(res, {"people": [{"name": "Iris"}]})
 
@@ -62,7 +62,7 @@ class TestWrap(TestBase):
             t = {
                 people: set({firstname: 'Iris'})
             };
-           return(t.wrap('People'), 2);
+           return t.wrap('People'), 2;
         ''')
         self.assertEqual(res, {"people": [{}]})
 
@@ -70,7 +70,7 @@ class TestWrap(TestBase):
             t = PS{
                 people: set({firstname: 'Iris'})
             };
-           return(t.wrap('People'), 2);
+           return t.wrap('People'), 2;
         ''')
         self.assertEqual(res, {"people": [{}]})
 
@@ -78,7 +78,7 @@ class TestWrap(TestBase):
             t = {
                 people: set(Person{name: 'Iris'})
             };
-            return(t.wrap('Other'), 2);
+            return t.wrap('Other'), 2;
         ''')
         self.assertEqual(res, {})
 
@@ -86,7 +86,7 @@ class TestWrap(TestBase):
             t = PS{
                 people: set(Person{name: 'Iris'})
             };
-            return(t.wrap('Other'), 2);
+            return t.wrap('Other'), 2;
         ''')
         self.assertEqual(res, {})
 
@@ -178,7 +178,7 @@ class TestWrap(TestBase):
 
             .notes = set(.note);
 
-            return(.notesSummary = .wrap('_NotesSummary'), 3);
+            return .notesSummary = .wrap('_NotesSummary'), 3;
         ''')
 
         ids = await client.query('[.id(), .note.id(), .iris.id()];')
@@ -226,7 +226,7 @@ class TestWrap(TestBase):
             }, true);
             set_type('_Pdeep', {
                 dict: 'any',
-                deep: || return({a: {b: 5}}, 2)
+                deep: || return {a: {b: 5}}, 2
             }, true);
         ''')
 
@@ -278,7 +278,7 @@ class TestWrap(TestBase):
                 name: 'iris',
                 messages: ['hi', 'hello', 'bye']
             };
-            return(p.wrap('_Pmore'), 2);
+            return p.wrap('_Pmore'), 2;
         ''')
 
         self.assertEqual(res, {