+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.2.3
 
 * Added `closure(..)` function to create closures from string, issue #273.
+* Fixed task replication bug in `set_closure(..)` function, issue #274.
 
 # v1.2.2