+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added `closure(..)` function to create closures from string, issue #273.
 * Fixed task replication bug in `set_closure(..)` function, issue #274.
+* Fixed task replication bug in `set_owner(..)` function, issue #275.
 
 # v1.2.2